Updated upload_archives by moving all the S3 code into s3_client
